tests, libstorage: relocate some recently introduced PVC utils
These were introduced in #7324 and would perfectly fit in libstorage.
